Move AWS gateway CDK files into gateways
