+++ .env.example
@@ -163,7 +163,7 @@ COOKIE_DOMAIN=localhost
 # =============================================================================
 # AWS Lambda environment (auto-injected by CDK — do NOT set manually)
 #
-# Source of truth: infra/aws/lib/api-gateway.ts (environment blocks)
+# Source of truth: infra/aws/lib/gateways/api-gateway.ts (environment blocks)
 #
 # CDK resolves these from CloudFormation refs at deploy time.
 # Listed here for reference when debugging a running Lambda.
