+++ infra/aws/lib/gateways/api-gateway-access-log.ts
