+++ infra/aws/lib/gateways/auth-gateway.test.ts
